+++ src/scripts/tasks.js
@@ -1,26 +1,7 @@
 const completed = (() => {
-  // const taskOne = {
-  //   description: 'Clean car',
-  //   completed: true,
-  //   index: 0,
-  // };
-  // const taskTwo = {
-  //   description: 'Do insurance',
-  //   completed: false,
-  //   index: 1,
-  // };
-  // const taskThree = {
-  //   description: 'Deposit money',
-  //   completed: false,
-  //   index: 2,
-  // };
-  // const tasksArray = [];
-  // tasksArray.push(taskOne);
-  // tasksArray.push(taskTwo);
-  // tasksArray.push(taskThree);
   let tasksArray = [];
   if (localStorage.getItem('tasksListArray') != null) {
-    tasksArray = JSON.parse(localStorage.getItem('tasksListArray'));
+     tasksArray = JSON.parse(localStorage.getItem('tasksListArray'));
   }
 
   tasksArray.forEach((task) => {
